photoshoot work

+++ badger_main/Models/ProductPhotoshoots.cs
@@ -21,7 +21,7 @@ namespace badgerApi.Models
     {
         public int product_shoot_status_id { get; set; }
         public int updated_by { get; set; }
-        public int updated_at { get; set; }
+        public double updated_at { get; set; }
         
     }
 }
